requirements.txt
APScheduler==3.6.3
async-timeout==3.0.1
attrs==21.2.0
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.4
idna==3.2
idna-ssl==1.1.0
multidict==5.1.0
Pillow==8.3.1
pyTelegramBotAPI==2.2.3
python-telegram-bot==13.7
pytz==2021.1
requests==2.26.0
six==1.16.0
tornado==6.1
typing-extensions==3.10.0.0
tzlocal==2.1
urllib3==1.26.6
yarl==1.6.3
